feat: extract TeamEntity into @gochamps/domain-types as extraction pilot
apps/cms/src/Teams/state.ts re-exports TeamEntity/CoachEntity as local
interfaces extending the package types (rather than a bare `export {} from`)
because isolatedModules requires `export type` for type-only re-exports,
and the app's pinned prettier@1.19.1 predates that TS 3.8 syntax and
cannot format it. All 28 existing consumers keep importing from
'../Teams/state' unmodified.
