+++ packages/domain-types/package.json
@@ -1,0 +1,7 @@
+{
+  "name": "@gochamps/domain-types",
+  "version": "0.0.0",
+  "private": true,
+  "main": "src/index.ts",
+  "types": "src/index.ts"
+}
